+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Set-PnPDefaultColumnValues: Fixed character encoding issue on folders #1706
 - Fixed import of search configuration to tenant via string
 - Set-PnPTenantSite: Added support for setting default sharing and sharing permissions
+- ConvertTo-PnPClientSidePage: Added support for copying page metadata to the modern version of the page + parameter to clear the transformation cache
 
 ## [3.5.1901.0]